requirements.txt
# Add your dependencies here
streamlit
reportlab
streamlit-drawable-canvas
pillow
numpy